fix: taxonomy returned null and parser only parsed molecules with activity values
